refac backend sur les mapper achvte/ligne/produits

+++ src/main/java/com/projet/equipement/dto/achat/AchatUpdateDto.java
@@ -17,7 +17,7 @@ public class AchatUpdateDto {
 
 //    private Integer clientId;
 
-    private Integer employeId;
+    private Long employeId;
 
     private Boolean actif;
